review model - project - configuration
